implmented Service layer
